Frontend models

+++ Backend API/ICT302-BackendAPI/ICT302-BackendAPI.Database/Repositories/JobsPendingRepository.cs
@@ -8,10 +8,13 @@ namespace ICT302_BackendAPI.Database.Repositories
 {
     public class JobsPendingRepository(SchemaContext ctx) : IJobsPendingRepository
     {
-        public async Task<List<JobsPending>> GetJobsPendingAsync()
+        public async Task<List<JobsPending>?> GetJobsPendingAsync()
         {
+            if (!await ctx.CheckDbIsAvailable())
+                return null;
+            
             var jobsPending = await ctx.JobsPending.ToListAsync();
-            if(jobsPending.Any())
+            if(jobsPending.Count != 0)
             {
                 jobsPending.ForEach(j => ctx.JobsPending.Attach(j));
             }
@@ -20,23 +23,21 @@ namespace ICT302_BackendAPI.Database.Repositories
 
         public async Task<JobsPending?> GetPendingJobByQueuePosition(int queuePosition)
         {
+            if (!await ctx.CheckDbIsAvailable())
+                return null;
+            
             var jobs = await ctx.JobsPending.ToListAsync();
             var job = jobs.Find(j => j.QueueNumber == queuePosition);
-            if (job != null)
-            {
-                await ctx.Entry(job).Reference(j => j.JobDetails).LoadAsync();
-                await ctx.Entry(job.JobDetails).Reference(j => j.Graphic).LoadAsync();
-                await ctx.Entry(job.JobDetails.Graphic).Reference(j => j.Animal).LoadAsync();
-                await ctx.Entry(job.JobDetails).Reference(j => j.Model3D).LoadAsync();
-            }
+            
+            if (job == null) return null;
+            
+            await ctx.Entry(job).Reference(j => j.JobDetails).LoadAsync();
+            await ctx.Entry(job.JobDetails).Reference(j => j.Graphic).LoadAsync();
+            await ctx.Entry(job.JobDetails.Graphic!).Reference(j => j!.Animal).LoadAsync();
+            await ctx.Entry(job.JobDetails).Reference(j => j.Model3D).LoadAsync();
+            
 
             return job;
-        }
-
-        public int DeleteJobsPending(JobsPending jobsPending)
-        {
-            ctx.JobsPending.Remove(jobsPending);
-            return ctx.SaveChanges();
         }
 
         public async Task<JobsPending?> GetJobsPendingByIDAsync(int id)
@@ -47,11 +48,19 @@ namespace ICT302_BackendAPI.Database.Repositories
             return job;
         }
 
-        public async Task<JobsPending> CreateJobsPendingAsync(JobsPending jobsPending)
+        public async Task<bool> CheckDbAvailability()
         {
+            return await ctx.CheckDbIsAvailable();
+        }
+
+        public async Task<JobsPending?> CreateJobsPendingAsync(JobsPending jobsPending)
+        {
+            if (!await ctx.CheckDbIsAvailable())
+                return null;
+            
             // New pending jobs objects will default to -1 queue number until created
             var currentPendingJobs = await GetJobsPendingAsync();
-            if (currentPendingJobs.Any())
+            if (currentPendingJobs != null && currentPendingJobs.Any())
             {
                 jobsPending.QueueNumber = currentPendingJobs.Count + 1;
             }
@@ -65,15 +74,21 @@ namespace ICT302_BackendAPI.Database.Repositories
             return jobsPending;
         }
 
-        public async Task<JobsPending> UpdateJobsPendingAsync(JobsPending jobsPending)
+        public async Task<JobsPending?> UpdateJobsPendingAsync(JobsPending jobsPending)
         {
+            if (!await ctx.CheckDbIsAvailable())
+                return null;
+            
             ctx.JobsPending.Update(jobsPending);
             await ctx.SaveChangesAsync();
             return jobsPending;
         }
 
-        public async Task<int> DeleteJobsPendingAsync(JobsPending jobsPending)
+        public async Task<int?> DeleteJobsPendingAsync(JobsPending jobsPending)
         {
+            if (!await ctx.CheckDbIsAvailable())
+                return null;
+            
             ctx.JobsPending.Remove(jobsPending);
             return await ctx.SaveChangesAsync();
         }
